Subforums should have unique names.
Zach and I already agreed on this in person, but we never put it in
code. Today during development, I ended up with five subforums with each
name. I can't reproduce that issue any more, but having this validation
(and associated change in seed data), will make sure that the next time
something wonky like this happens, we'll know about it.