Add AGPL3 LICENSE file
The readme states that the code in this repository is licensed under the
AGPL, version 3. Add a LICENSE file so that the license is available in
the repository itself, and so that GitHub is able to understand the 
license, which allows it to show the appropriate UI and include the repo
in searches by license.